# try-go-openapi/server/go/api_default.go
/*
 * 匿名掲示板API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// CreateComment - 
func CreateComment(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// CreatePost - 
func CreatePost(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// DeleteComment - 
func DeleteComment(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// DeletePost - 
func DeletePost(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetComments - 
func GetComments(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetPost - 
func GetPost(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetPosts - 
func GetPosts(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// UpdatePost - 
func UpdatePost(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}