# try-go-openapi/server/go/model_comment_request.go
/*
 * 匿名掲示板API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type CommentRequest struct {

	Comment *CommentParameters `json:"comment"`
}