# try-go-openapi/server/go/model_error_properties.go
/*
 * 匿名掲示板API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type ErrorProperties struct {

	Title string `json:"title"`

	Detail string `json:"detail"`
}