# try-go-openapi/server/go/model_post_properties.go
/*
 * 匿名掲示板API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"time"
)

type PostProperties struct {

	Id int32 `json:"id"`

	Title string `json:"title"`

	Content string `json:"content"`

	PostedAt time.Time `json:"posted_at"`
}